+++ photon-targeting/src/test/java/org/photonvision/targeting/PhotonPipelineResultTest.java
@@ -182,7 +182,7 @@ public class PhotonPipelineResultTest {
                         new MultiTargetPNPResult(
                                 new PNPResult(
                                         new Transform3d(new Translation3d(1, 2, 3), new Rotation3d(1, 2, 3)), 0.1),
-                                List.of(1, 2, 3)));
+                                List.of((short) 1, (short) 2, (short) 3)));
         b =
                 new PhotonPipelineResult(
                         3,
@@ -234,7 +234,7 @@ public class PhotonPipelineResultTest {
                         new MultiTargetPNPResult(
                                 new PNPResult(
                                         new Transform3d(new Translation3d(1, 2, 3), new Rotation3d(1, 2, 3)), 0.1),
-                                List.of(1, 2, 3)));
+                                List.of((short) 1, (short) 2, (short) 3)));
         assertEquals(a, b);
     }
 
@@ -389,7 +389,7 @@ public class PhotonPipelineResultTest {
                         new MultiTargetPNPResult(
                                 new PNPResult(
                                         new Transform3d(new Translation3d(1, 8, 3), new Rotation3d(1, 2, 3)), 0.1),
-                                List.of(3, 4, 7)));
+                                List.of((short) 3, (short) 4, (short) 7)));
         b =
                 new PhotonPipelineResult(
                         3,
@@ -441,7 +441,7 @@ public class PhotonPipelineResultTest {
                         new MultiTargetPNPResult(
                                 new PNPResult(
                                         new Transform3d(new Translation3d(1, 2, 3), new Rotation3d(1, 2, 3)), 0.1),
-                                List.of(1, 2, 3)));
+                                List.of((short) 1, (short) 2, (short) 3)));
         assertNotEquals(a, b);
     }
 }
