+++ photon-targeting/src/test/java/org/photonvision/targeting/proto/MultiTargetPNPResultProtoTest.java
@@ -40,7 +40,7 @@ public class MultiTargetPNPResultProtoTest {
                 new MultiTargetPNPResult(
                         new PNPResult(
                                 new Transform3d(new Translation3d(1, 2, 3), new Rotation3d(1, 2, 3)), 0.1),
-                        List.of(1, 2, 3));
+                        List.of((short) 1, (short) 2, (short) 3));
         serializedResult = MultiTargetPNPResult.proto.createMessage();
         MultiTargetPNPResult.proto.pack(serializedResult, result);
         unpackedResult = MultiTargetPNPResult.proto.unpack(serializedResult);
