+++ photon-targeting/src/test/java/org/photonvision/targeting/proto/PhotonPipelineResultProtoTest.java
@@ -142,7 +142,7 @@ public class PhotonPipelineResultProtoTest {
                         new MultiTargetPNPResult(
                                 new PNPResult(
                                         new Transform3d(new Translation3d(1, 2, 3), new Rotation3d(1, 2, 3)), 0.1),
-                                List.of(1, 2, 3)));
+                                List.of((short) 1, (short) 2, (short) 3)));
         serializedResult = PhotonPipelineResult.proto.createMessage();
         PhotonPipelineResult.proto.pack(serializedResult, result);
         unpackedResult = PhotonPipelineResult.proto.unpack(serializedResult);
